+++ src/node-lib/curriculum-api-2023/queries/teachersUnitOverview/teachersUnitOverview.query.test.ts
@@ -1,6 +1,7 @@
 import {
   syntheticUnitvariantLessonsByKsFixture,
   lessonDataFixture,
+  programmeFieldsFixture,
   lessonFixture,
 } from "@oaknational/oak-curriculum-schema";
 
@@ -71,6 +72,13 @@ const unitPageFixture2 = syntheticUnitvariantLessonsByKsFixture({
   },
 });
 
+export const unitsInOtherProgrammesFixture = [
+  {
+    programme_slug: "programme-slug",
+    programme_fields: programmeFieldsFixture(),
+  },
+];
+
 export const unitSequenceFixture: UnitSequence = [
   {
     unitSlug: "unit-1",
@@ -111,6 +119,7 @@ describe("teachersUnitOverview", () => {
           Promise.resolve({
             lessons: [],
             unitSequence: unitSequenceFixture,
+            unitsInOtherProgrammes: unitsInOtherProgrammesFixture,
           }),
         ),
       })({
@@ -126,6 +135,7 @@ describe("teachersUnitOverview", () => {
         Promise.resolve({
           lessons: [syntheticUnitvariantLessonsByKsFixture()],
           unitSequence: unitSequenceFixture,
+          unitsInOtherProgrammes: unitsInOtherProgrammesFixture,
         }),
       ),
     })({
@@ -180,6 +190,8 @@ describe("teachersUnitOverview", () => {
         title: "Unit 2",
       },
       prevUnit: null,
+      tierOptionToggles: [],
+      subjectOptionToggles: [],
     });
   });
   it("returns lessons in the correct order", async () => {
@@ -189,6 +201,7 @@ describe("teachersUnitOverview", () => {
         Promise.resolve({
           lessons: [unitPageFixture2, unitPageFixture],
           unitSequence: unitSequenceFixture,
+          unitsInOtherProgrammes: unitsInOtherProgrammesFixture,
         }),
       ),
     })({
@@ -219,6 +232,7 @@ describe("teachersUnitOverview", () => {
               },
             ],
             unitSequence: unitSequenceFixture,
+            unitsInOtherProgrammes: unitsInOtherProgrammesFixture,
           }),
         ),
       })({