+++ src/node-lib/curriculum-api-2023/queries/topNav/topNav.query.ts
@@ -15,6 +15,7 @@ import OakError from "@/errors/OakError";
 
 const topNavQuery = (sdk: Sdk) => async (): Promise<TopNavProps> => {
   const res = await sdk.topNav();
+
   const parsed = topNavResponseSchema.safeParse(res);
 
   if (!parsed?.success) {